added enable/disable Behavior

+++ src/PushBehavior.php
@@ -2,11 +2,11 @@
 
 namespace lav45\behaviors;
 
+use lav45\behaviors\traits\WatchAttributesTrait;
 use yii\base\Behavior;
 use yii\db\ActiveRecord;
 use yii\db\ActiveRecordInterface;
 use yii\db\AfterSaveEvent;
-use lav45\behaviors\traits\WatchAttributesTrait;
 
 /**
  * Class PushBehavior
@@ -18,6 +18,9 @@ use lav45\behaviors\traits\WatchAttributesTrait;
  *          [
  *              'class' => PushBehavior::class,
  *              'relation' => 'apiUser',
+ *              'enable' => static function () {
+ *                  return true;
+ *              },
  *              'attributes' => [
  *                  // Observe the change in the `status` attribute
  *                  // Writes the "value" in field `status` the relation model
@@ -56,6 +59,15 @@ use lav45\behaviors\traits\WatchAttributesTrait;
 class PushBehavior extends Behavior
 {
     use WatchAttributesTrait;
+
+    /**
+     * @var bool|\Closure
+     * Can be passed to \Closure for enable or disable Behavior
+     * function () {
+     *      return true;
+     * }
+     */
+    public $enable = true;
 
     /**
      * @var string target relation name
@@ -109,6 +121,10 @@ class PushBehavior extends Behavior
      */
     final public function afterInsert()
     {
+        if ($this->isEnable() === false) {
+            return;
+        }
+
         foreach ($this->getRelationIterator() as $model) {
             if (null === $model) {
                 if (false === $this->createRelation) {
@@ -135,6 +151,10 @@ class PushBehavior extends Behavior
      */
     final public function afterUpdate(AfterSaveEvent $event)
     {
+        if ($this->isEnable() === false) {
+            return;
+        }
+
         if ($changedAttributes = $this->getChangedAttributes($event->changedAttributes)) {
             foreach ($this->getRelationIterator(true) as $model) {
                 $this->updateModel($model, $changedAttributes);
@@ -153,6 +173,10 @@ class PushBehavior extends Behavior
      */
     final public function beforeDelete()
     {
+        if ($this->isEnable() === false) {
+            return;
+        }
+
         foreach ($this->getRelationIterator(true) as $model) {
             if (true === $this->deleteRelation) {
                 $model->delete();
@@ -193,5 +217,16 @@ class PushBehavior extends Behavior
                 yield $item;
             }
         }
+    }
+
+    /**
+     * @return bool
+     */
+    protected function isEnable()
+    {
+        if (is_bool($this->enable)) {
+            return $this->enable;
+        }
+        return (bool)call_user_func($this->enable);
     }
 }
